Move playersState and related from server_common.go to logic.go.
